+++ R/create_model_plots.R
@@ -10,9 +10,17 @@ create_model_plots <- function(tests, model)
   # Read the rain summary from the attribute "rain_summary"
   rain_summary <- kwb.utils::getAttribute(data_overview_plot, "rain_summary")
 
+  create_plot <- function(gg_plot, title, description) {
+    list(
+      gg_plot = gg_plot,
+      title = title,
+      description = description
+    )
+  }
+
   list(
     # Plot 1
-    list(
+    create_plot(
       gg_plot = plot_model_calibration(tests, ref_model),
       title = "\u00dcberblick \u00fcber getestete Modelle",
       description = paste(
@@ -21,7 +29,7 @@ create_model_plots <- function(tests, model)
       )
     ),
     # Plot 2
-    list(
+    create_plot(
       gg_plot = data_overview_plot,
       title = "Datenqualit\u00e4t und Kalibrationsbereich",
       description = paste(
@@ -32,7 +40,7 @@ create_model_plots <- function(tests, model)
       )
     ),
     # Plot 3
-    list(
+    create_plot(
       gg_plot = plot_fit_vs_data(model),
       title = "Modelbewertung gegen Messwerte",
       description = paste(
@@ -42,29 +50,29 @@ create_model_plots <- function(tests, model)
       )
     ),
     # Plot 4
-    list(
+    create_plot(
       gg_plot = plot_contigency(model),
       title = "Perzentile auf Basis klassifizierter Datenpunkte",
       description = paste(
         "Berechnete Perzentile (y-Achse) sollten mit Klassifizierung x-Achse ",
         "\u00fcbereinstimmen"
       )
-    ),
-    # Plot 5
-    list(
-      gg_plot = plot_effects(model),
-      title = "Modellimplikationen",
-      description = paste(
-        "Das sind die Modellimplikationen..."
-      )
-    ),
-    # Plot 6
-    list(
-      gg_plot = empty_ggplot("Leere Grafik."),
-      title = "Grafik 6: Noch nicht vergeben",
-      description = paste(
-        "Keine Grafik - keine Beschreibung."
-      )
     )
+    # # Plot 5
+    # , create_plot(
+    #   gg_plot = plot_effects(model),
+    #   title = "Modellimplikationen",
+    #   description = paste(
+    #     "Das sind die Modellimplikationen..."
+    #   )
+    # )
+    # # Plot 6
+    # , create_plot(
+    #   gg_plot = empty_ggplot("Leere Grafik."),
+    #   title = "Grafik 6: Noch nicht vergeben",
+    #   description = paste(
+    #     "Keine Grafik - keine Beschreibung."
+    #   )
+    # )
   )
 }